LayoutTests/media/loopstart-limits.html


<video loopstart=1.5></video>
<audio loopstart=1.5></audio>

<video start=1.0></video>
<audio start=1.0></audio>

<video ></video>
<audio ></audio>

<video loopstart="s1.0"></video>
<audio loopstart="s1.0"></audio>

<video loopstart='s1.0' start=1.0></video>
<audio loopstart='s1.0' start=1.0></audio>

<p>Test loopstart attribute parsing, limits, and defaults.</p>

<script src=video-test.js></script>

<script>

// 4.8.10.5. Offsets into the media resource
// loopstart: default the value of the start DOM attribute
// start: default value is 0

consoleWrite("<em>&lt;video loopstart=1.5&gt;</em> - loopstart specified, should not change");

var video = document.getElementsByTagName('video')[0];
var audio = document.getElementsByTagName('audio')[0];
testExpected("video.loopStart", 1.5);
testExpected("audio.loopStart", 1.5);

consoleWrite("<br><em>&lt;video start=1.0&gt;</em> - loopstart not specified, should return start attr");
var video = document.getElementsByTagName('video')[1];
var audio = document.getElementsByTagName('audio')[1];
testExpected("video.loopStart", 1.0);
testExpected("audio.loopStart", 1.0);

consoleWrite("<br><em>&lt;video&gt;</em> - no loopstart and no start, default should be 0");
var video = document.getElementsByTagName('video')[2];
var audio = document.getElementsByTagName('audio')[2];
testExpected("video.loopStart", 0);
testExpected("audio.loopStart", 0);

consoleWrite("<br><em>&lt;video loopstart='s1.0'&gt;</em> - bogus loopstart and no start, default should be 0 (start default)");
var video = document.getElementsByTagName('video')[3];
var audio = document.getElementsByTagName('audio')[3];
testExpected("video.loopStart", 0);
testExpected("audio.loopStart", 0);

consoleWrite("<br><em>&lt;video loopstart='s1.0' start=1.0&gt;</em> - bogus loopstart, default should be start attr");
var video = document.getElementsByTagName('video')[4];
var audio = document.getElementsByTagName('audio')[4];
testExpected("video.loopStart", 1.0);
testExpected("audio.loopStart", 1.0);


endTest()

</script>
